Minor DrawSprite fixes
- Updated UiModule::Utils::DrawSprite functions
    - Updated one of function overloads to take sprite width and height as int
    - Fixed the x1 y1 x2 y2 overload UV mapping

+++ src/modules/ui/utils/draw-sprite.h
@@ -23,8 +23,8 @@ namespace UiModule::Utils {
 		uint16_t remap,
 		float centerX,
 		float centerY,
-		float spriteWidth,
-		float spriteHeight,
+		int spriteWidth,
+		int spriteHeight,
 		float scale,
 		uint8_t alpha = 0xFF
 	);
